DurationToIntegerConverter: dumb down to use long instead of BigInteger
The old Hibernate MySQL 5 dialect used by the application does not
properly support converters for basic fields that are of type
BigInteger, even though that type is recognized by the JPA
specification.

Therefore, dumb down the converter to use a long, instead of a integer
with arbitrary precision. This comes at the cost that durations longer
than 2^63 - 1 nanoseconds (292.47 years) can no longer be represented.
This should be an acceptable compromise in practice, given the
inconvenience of storing raw blobs more directly.