Add unit tests and integration tests with junit and Arquillian
These tests are meant to showcase the suitability of the designed
environment for carrying out the good testing practices studied in the
subject.

The PBKDF2 password hashing module was removed because it caused
problems during the deployment of the reduced Arquillian web archive, as
I could not come up with a way to make the archive register a new JBoss
module when it is deployed.

The Arquillian REST test extension has trouble working because it can't
get the address where Thorntail is listening to, but it doesn't do much
anyway in our case, so let's just not use it.