Add GitHub Actions CI and CD workflow
In addition to adding the workflow itself, a standard Maven wrapper was
added to ensure that both developers and the workflow use the same Maven
version, to prevent unexpected discrepancies.

The POM of the web project was modified a bit to allow configuration of
where the Docker image is deployed via environment variables.