chore: update Docker compose file to point to the new repository
The Docker repository was changed in order to better decouple it from
internal Maven identifiers, and allow for a different staging
environment whose only difference with the production environment is
that the deployment occurs to another Docker Hub repository.

The previously stated changes made this Docker compose file no longer
function, so update it accordingly.